% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSimFun.R
\name{FindSimFun}
\alias{FindSimFun}
\title{Sample-cell line pathway activity-based similarity score}
\usage{
FindSimFun(
  Patient.D.Genes,
  Patient.Reactome,
  Patient.Expr.Meth.CNV,
  Cell.D.Genes,
  Cell.Reactome,
  Cell.Expr.Meth.CNV
)
}
\arguments{
\item{Patient.D.Genes}{A vector of DE or DM or DA genes of a tumor sample}

\item{Patient.Reactome}{A R dataframe for patient-specific enriched biological pathways with three columns labeled as ID, reactome_pathway, and p.adjust.pat, respectively.}

\item{Patient.Expr.Meth.CNV}{A R dataframe with sample-specific deconvoluted gene expression or deconvoluted DNA methylation or copy number profiles.}

\item{Cell.D.Genes}{A vector of DE or DM or DA genes of a cell line}

\item{Cell.Reactome}{A R dataframe for cell line-specific enriched biological pathways with three columns labeled as ID, reactome_pathway, and p.adjust.pat, respectively.}

\item{Cell.Expr.Meth.CNV}{A R dataframe with cell line-specific gene expression or DNA methylation or copy number profiles.}
}
\value{
A numeric value (Spearman similarity score)
}
\description{
This function computes Spearman rank correlation between each sample-cell line pair using the sample-specific deconvoluted expression, sample-specific deconvoluted DNA methylation, and gene-centric copy number values of samples. All DM genes that occur in an enriched pathway of a sample or cell line are used to compute the Spearman rank correlation between each sample-cell line pair. Similarly, using DE and DA genes that are present in the gene list of the union of enriched pathways, it computes the Spearman rank correlation between each sample-cell line pair. Using DM and DE genes, it uses the deconvoluted profiles of samples to compute Spearman rank correlation whereas for DA genes, it uses the gene-centric copy number profile of samples to compute Spearman rank correlation.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 5 | Function 1 |
}
