% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellLineDAFun.R
\name{GetCellLineDAFun}
\alias{GetCellLineDAFun}
\title{Cell line-specific differentially aberrated (DA) genes}
\usage{
GetCellLineDAFun(CNV.CellLine, ncores = 1, CTDDirectory = "~")
}
\arguments{
\item{CNV.CellLine}{A R dataframe with gene-centric copy number values of cell lines. Rows are the genes, columns are the cell lines.}

\item{ncores}{An integer value for the number of cores.}

\item{CTDDirectory}{A character string for the file path of the directory for the output to be stored.}
}
\value{
R dataframe objects each with a single column containing DA genes for each cell line. The column is labeled as gene.
}
\description{
This function selects the highly amplified and deleted genes for each cell line compared with all other cell lines from various cancer types to capture the cancer-specific variation in copy number profiles in computing the similarity score between a patient sample and cell line. It selects the cell line-specific DA genes based on Tukey’s mean-difference curve based on gene-centric copy number data of cell lines. To determine whether a gene is DA for a specific cell line, it calculates the mean of that gene's copy number values across all cell lines and a difference between the copy number value of that gene in the cell line under consideration and the mean of the copy number values of that gene in the other cell lines. For each cell line, the highly amplified genes are chosen based on mean ≥ 0.5 and difference ≤ 1, and highly deleted genes are chosen based on mean ≥ 0.5 and difference ≤ -1.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 4 | Function 1 |
}
