% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellLineDEFun.R
\name{GetCellLineDEFun}
\alias{GetCellLineDEFun}
\title{Getting cell line-specific differentially expressed (DE) genes}
\usage{
GetCellLineDEFun(
  RnaSeq_data,
  parallel = c("TRUE", "FALSE"),
  ncores = 2,
  CTDDirectory = "~"
)
}
\arguments{
\item{RnaSeq_data}{A R dataframe with gene-centric expression values of cell lines. Rows are the genes, columns are the tumor samples.}

\item{parallel}{A boolean value ('TRUE' or 'FALSE')indicating if the users want to run this function in multi-core.}

\item{ncores}{A integer value for the number of cores.}

\item{CTDDirectory}{A character string for the file path of the directory for the output to be stored.}
}
\value{
R dataframe objects each with a single column containing DE genes for each cell line. The column is labeled as gene.
}
\description{
This function computes the median expression of a gene across all cell lines and further computed the fold change of that gene to the computed median expression. A gene was considered up if the fold-change ≥ 4 and down if the fold-change ≤ −4.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 3 | Function 2 |
}
