% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellLineDMFun.R
\name{GetCellLineDMFun}
\alias{GetCellLineDMFun}
\title{Cell line-specific differentially (DM) genes}
\usage{
GetCellLineDMFun(
  DNAmethylation_data,
  parallel = c("TRUE", "FALSE"),
  ncores = 2,
  CTDDirectory = "~"
)
}
\arguments{
\item{DNAmethylation_data}{A R dataframe with gene-centric DNA methylation beta values of cell lines. Rows are the genes, columns are the cell lines.}

\item{parallel}{A boolean value ('TRUE' or 'FALSE')indicating if the users want to run this function in multi-core.}

\item{ncores}{A integer value for the number of cores.}

\item{CTDDirectory}{A character string for the file path of the directory for the output to be stored.}
}
\value{
R dataframe objects each with a single column containing DM genes for each cell line. The column is labeled as gene.
}
\description{
This function computes cell line-specific DM genes for each cell lines It computes M-values from gene-centric beta values and computes the median of gene-centric M-values across all the cell lines It considers a gene hypermethylated if the M-value fold-change ≥ 4 and hypomethylated if fold-change ≤ -4.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 4 | Function 2 |
}
