% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPathFun.R
\name{GetPathFun}
\alias{GetPathFun}
\title{Enriched biological pathways}
\usage{
GetPathFun(CancerGeneList, SampleCell_de_genes)
}
\arguments{
\item{CancerGeneList}{A R dataframe with a column of frequently mutated cancer driver genes.}

\item{SampleCell_de_genes}{A R dataframe with a column containing DE genes of a sample or cell line.}
}
\value{
An R dataframe object with three columns labeled as 'ID', 'reactome_pathway', and 'p.adjust.pat', respectively.
}
\description{
This function identifies the enriched biological pathways for each sample and cell line utilizing the DE genes. It finds enriched biological pathways listed in the REACTOME database via a pathway enrichment tool Pathfinder. To prioritize cancer-related biological pathways, specifically, it uses the known frequently mutated genes that were considered cancer-driving genes from the DE gene list for this analysis. We obtained the frequently mutated cancer-driving genes from the COSMIC database. GetPath() considers the pathways that were significantly enriched with FDR-corrected hypergeometric p-values < 0.05 in this DE cancer gene list.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 3 | Function 3 |
}
