% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleDMFun.R
\name{GetSampleDMFun}
\alias{GetSampleDMFun}
\title{Sample-specific differentially methylated (DM) genes}
\usage{
GetSampleDMFun(
  DNAmethylation_data,
  parallel = c("TRUE", "FALSE"),
  ncores = 2,
  CTDDirectory = "~"
)
}
\arguments{
\item{DNAmethylation_data}{A R dataframe with gene-centric DNA methylation beta values of bulk tumor samples. Rows are the genes, columns are the tumor samples.}

\item{parallel}{A boolean value ('TRUE' or 'FALSE')indicating if the users want to run this function in multi-core.}

\item{ncores}{An integer value for the number of cores.}

\item{CTDDirectory}{A character string for the file path of the directory for the output to be stored.}
}
\value{
R dataframe objects each with a single column containing DM genes for each tumor sample. The column is labeled as gene.
}
\description{
This function computes sample-specific DM genes for each sample. It computes M-values from gene-centric beta values and computes the median of gene-centric M-values across all the samples. It considers a gene hypermethylated if the M-value fold-change ≥ 4 and hypomethylated if fold-change ≤ -4.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 4 | Function 1 |
}
