% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_cnv.R
\docType{data}
\name{pat_cnv}
\alias{pat_cnv}
\title{Copy number data of patient tumor samples}
\format{
An object of class \code{data.frame} with 2000 rows and 1 columns.
}
\usage{
pat_cnv
}
\description{
Copy number data of patient tumor samples
}
\keyword{datasets}
