% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDeconvMethylFun.R
\name{PlotDeconvMethylFun}
\alias{PlotDeconvMethylFun}
\title{Correlation plot to show the correlation between the reference cell types and estimated clusters DNA methylation deconvolution}
\usage{
PlotDeconvMethylFun(
  Methylation.Sample,
  Reference.Methylation.Probes,
  Reference.Methylation.CellTypes,
  stage1_result_ct,
  Reference.CellTypes.Names
)
}
\arguments{
\item{Methylation.Sample}{A R dataframe for the DNA methylation values for the samples and the probes.}

\item{Reference.Methylation.Probes}{A vector of the marker loci (probes) based on comparisons of each class of reference against all other samples using t-test.}

\item{Reference.Methylation.CellTypes}{A R dataframe for the DNA methylation values for the reference cell types and the probes.}

\item{stage1_result_ct}{The output from RunDeconvMethylFun()}

\item{Reference.CellTypes.Names}{A vector of the names of the reference cell types.}
}
\value{
Correlation Plot.
}
\description{
This function creates a correlation plot to show the correlation between the reference cell types used and estimated clusters after DNA methylation deconvolution. Users can use this plot to choose the appropriate labels for the cell types for the deconvoluted DNA methylation profiles.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 1 | Function 2 |
}
