% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbeToGeneFun.R
\name{ProbeToGeneFun}
\alias{ProbeToGeneFun}
\title{Converting DNA methylation data from probes to genes}
\usage{
ProbeToGeneFun(
  Methylation.Probe.Annotation,
  Deconv_methylation,
  Annotation.skim
)
}
\arguments{
\item{Methylation.Probe.Annotation}{A R dataframe object with 'probe', 'symbol' and 'synonym' columns for 27K or 'ID', 'UCSC_REFGENE_NAME', 'UCSC_REFGENE_GROUP' columns for 450K DNA methylation data according to the convention of the annotation file in "Genome Browser" (https://genome.ucsc.edu/.}

\item{Deconv_methylation}{A R dataframe with DNA methylation values for each probe and cell type.}

\item{Annotation.skim}{A character string indicating if the user is using either 450K or 27K probes.}
}
\value{
An R dataframe object with deconvoluted DNA methylation values for each gene in cell types. The rows are the genes and the columns are the cell types. This function labels the columns as V1, V2, V3,....etc.
}
\description{
This function converts the DNA methylation data from probe-centric to gene-centric. This function works for both the 450K probe format and 27K probe format.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 1 | Function 3 |
}
