% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDeconvExprFun.R
\name{RunDeconvExprFun}
\alias{RunDeconvExprFun}
\title{Deconvolution of gene expression of bulk tumor samples}
\usage{
RunDeconvExprFun(Expression.Sample, Deconv_proportions)
}
\arguments{
\item{Expression.Sample}{A R dataframe for the gene-centric gene expression (RNASeq) values of the tumor samples.}

\item{Deconv_proportions}{A R dataframe with estimated cell proportions in tumor samples from output of the RunDeconvMethylFun() function.}
}
\value{
A R dataframe with gene expression values in constituent cell types. The rows are the genes and the columns are the cell types. This function labels the columns as \link{,1}, \link{,2},\link{,3},....etc.
}
\description{
This function utilizes the estimated cell proportions from Step 1 as a fixed input and computes an average gene expression profiles of constituent cell types through a constrained least-squares using quadratic programming.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 2 | Function 1 |
}
