% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleExprFun.R
\name{SampleExprFun}
\alias{SampleExprFun}
\title{Sample-specific deconvoluted gene expression profile.}
\usage{
SampleExprFun(Deconv_expression, Deconv_proportions, CTDDirectory = "~")
}
\arguments{
\item{Deconv_expression}{A R dataframe with the deconvoluted gene expression values for each gene in constituent cell types.}

\item{Deconv_proportions}{A R dataframe with the estimated proportions of constituent cell types in samples.}

\item{CTDDirectory}{A character string for the file path of the directory for the output to be stored.}
}
\value{
R dataframes including the deconvoluted gene expression profile for each sample in the specified directory. The rows are the genes and the columns are the cell types.This function labels the columns as V1, V2, V3,....etc.
}
\description{
This is the place for description.
}
\details{
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 2 | Function 2 |
}
