% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMethylFun.R
\name{SampleMethylFun}
\alias{SampleMethylFun}
\title{Sample-specific deconvoluted DNA methylation profile.}
\usage{
SampleMethylFun(Deconv_meth_gene, Deconv_proportions, CTDDirectory = "~")
}
\arguments{
\item{Deconv_meth_gene}{A R dataframe with the deconvoluted DNA methylation values for each gene in constituent cell types.}

\item{Deconv_proportions}{A R dataframe with the estimated proportions of constituent cell types in samples.}

\item{CTDDirectory}{A character string for the file path of the directory for the output to be stored.}
}
\value{
R dataframes including the deconvoluted DNA methylation profile for each sample in the specified directory. The rows are the genes and the columns are the cell types.This function labels the columns as V1, V2, V3,....etc.
}
\description{
This function utilizes the estimated proportion of constituent cell types and the estimated gene-centric DNA methylation profile of constituent cell type to compute the sample-specific deconvoluted DNA methylation profile for each tumor sample.
Created By: Banabithi Bose| Date Created: 4/10/2022 | Stage 1 | Function 4 |
}
