% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edec_stage_0.R
\name{tTestWithErrorHandling}
\alias{tTestWithErrorHandling}
\title{This chunk of code is from EDec paper
Select informative loci (EDec stage 0)}
\usage{
tTestWithErrorHandling(x)
}
\description{
\code{run_edec_stage_0} selects loci/probes that display cell type specific
patterns of methylation, and are likely to be informative for deconvolution.
This selection is based on comparison of reference methylation profiles
representing different cell types.
}
\details{
There are two versions of this method. The user can choose which one to use
by setting the \code{version} argument to either "one.vs.rest" (default) or
"each.pair". In the first version, the method will perform two-sample T-tests
to compare the levels of methylation on each locus/probe between references
of one class against all other references. These comparisons will be done for
each reference class in turn. The method will then select the most hyper- and
hypo-methylated loci/probes that showed p-value less than or equal to
\code{max_p_value}. The same number of markers will be selected from each
comparison, adding up to \code{num_markers}. The difference between the
"one.vs.rest" version and the "each.pair" version of the method is that,
instead of comparing one class of references against all others as is done in
"one.vs.rest" version, the "each.pair" version will perform comparisons
between each pair of reference classes.

p_1, data_group_2)
}
